+++ src/main/java/com/paneedah/mwc/creativetab/PropsTab.java
@@ -1,10 +1,13 @@
 package com.paneedah.mwc.creativetab;
 
-import com.paneedah.mwc.weapons.Guns;
 import net.minecraft.creativetab.CreativeTabs;
+import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
+import net.minecraft.util.ResourceLocation;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
+
+import static com.paneedah.mwc.utils.ModReference.ID;
 
 public class PropsTab extends CreativeTabs {
 	public PropsTab(int par1, String par2Str) {
@@ -14,6 +17,6 @@ public class PropsTab extends CreativeTabs {
 	@Override
 	@SideOnly(Side.CLIENT)
 	public ItemStack createIcon() {
-		return new ItemStack(Guns.AK12);
+		return new ItemStack(Item.REGISTRY.getObject(new ResourceLocation(ID, "barrel")));
 	}
 }